vector_db/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
google-generativeai
faiss-cpu
langchain